Began TerminalGame developement
